+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Drop monitoring of non-managed kiam pods.
+
 ## [0.2.0] - 2020-09-02
 
 - Add kube-proxy metrics